+++ src/objectedit.cpp
@@ -340,16 +340,16 @@ int cProcessManagement::objectedit_process_events(SDL_Event &event)
 					cWindow* w = new cObjectWindow(&Graphics.WM.texture, &Graphics.WM.font);
 					if (menuitem != NULL)
 						w->objects["objectmenu"]->SetText(0,menuitem->data2);
-					w->objects["posx"]->SetInt(3,(int)&o->pos.x);
-					w->objects["posy"]->SetInt(3,(int)&o->pos.y);
-					w->objects["posz"]->SetInt(3,(int)&o->pos.z);
-					w->objects["rotx"]->SetInt(3,(int)&o->rot.x);
-					w->objects["roty"]->SetInt(3,(int)&o->rot.y);
-					w->objects["rotz"]->SetInt(3,(int)&o->rot.z);
-					w->objects["scalex"]->SetInt(3,(int)&o->scale.x);
-					w->objects["scaley"]->SetInt(3,(int)&o->scale.y);
-					w->objects["scalez"]->SetInt(3,(int)&o->scale.z);
-					w->objects["objectname"]->SetInt(3, (int)&o->name);
+					w->objects["posx"]->SetInt(3,(intptr_t)&o->pos.x);
+					w->objects["posy"]->SetInt(3,(intptr_t)&o->pos.y);
+					w->objects["posz"]->SetInt(3,(intptr_t)&o->pos.z);
+					w->objects["rotx"]->SetInt(3,(intptr_t)&o->rot.x);
+					w->objects["roty"]->SetInt(3,(intptr_t)&o->rot.y);
+					w->objects["rotz"]->SetInt(3,(intptr_t)&o->rot.z);
+					w->objects["scalex"]->SetInt(3,(intptr_t)&o->scale.x);
+					w->objects["scaley"]->SetInt(3,(intptr_t)&o->scale.y);
+					w->objects["scalez"]->SetInt(3,(intptr_t)&o->scale.z);
+					w->objects["objectname"]->SetInt(3, (intptr_t)&o->name);
 					((cObjectWindow*)w)->undo = new cUndoChangeObject(Graphics.selectedobject);
 
 					Graphics.WM.addwindow(w);
