+++ src/wm/effectwindow.h
@@ -116,7 +116,7 @@ public:
 	cUndoItem* undo;
 	void* userfunc(void* param)
 	{
-		int p = (int)param;
+		intptr_t p = (intptr_t)param;
 		if(p == 0) // cancel
 		{
 			undo->undo();
