+++ src/wm/windowlistbox.h
@@ -39,8 +39,8 @@ public:
 
 	void drag();
 	void doubleclick();
-	void SetInt(int, int);
-	int GetInt(int);
+	void SetInt(int, intptr_t);
+	intptr_t GetInt(int);
 	string GetText(int);
 };
